# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Construction Coordination Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/proposals/{proposal_id}/approve": {
      "post": {
        "tags": [
          "runs-actions"
        ],
        "summary": "Approve",
        "operationId": "approve_api_proposals__proposal_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approval"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/proposals/{proposal_id}/execute": {
      "post": {
        "tags": [
          "runs-actions"
        ],
        "summary": "Execute",
        "operationId": "execute_api_proposals__proposal_id__execute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/operations/{operation_id}": {
      "get": {
        "tags": [
          "runs-actions"
        ],
        "summary": "Execution Receipt",
        "operationId": "execution_receipt_api_operations__operation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionExecution"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionExecution": {
        "properties": {
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "before_version": {
            "type": "integer",
            "title": "Before Version"
          },
          "after_version": {
            "type": "integer",
            "title": "After Version"
          },
          "status": {
            "type": "string",
            "const": "VERIFIED",
            "title": "Status",
            "default": "VERIFIED"
          },
          "mode": {
            "type": "string",
            "enum": [
              "simulated",
              "external"
            ],
            "title": "Mode"
          },
          "principal_id": {
            "type": "string",
            "title": "Principal Id"
          },
          "result": {
            "type": "string",
            "title": "Result"
          },
          "executed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Executed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "proposal_id",
          "project_id",
          "before_version",
          "after_version",
          "status",
          "mode",
          "principal_id",
          "result",
          "executed_at"
        ],
        "title": "ActionExecution"
      },
...
      "Approval": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "proposal_id": {
            "type": "string",
            "title": "Proposal Id"
          },
          "principal_id": {
            "type": "string",
            "title": "Principal Id"
          },
          "principal_role": {
            "type": "string",
            "title": "Principal Role"
          },
          "level": {
            "type": "string",
            "enum": [
              "standard",
              "strong"
            ],
            "title": "Level"
          },
          "confirmation": {
            "type": "string",
            "title": "Confirmation",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "proposal_id",
          "principal_id",
          "principal_role",
          "level",
          "confirmation",
          "created_at"
        ],
        "title": "Approval"
      },
      "ApprovalRequest": {
        "properties": {
          "strong": {
            "type": "boolean",
            "title": "Strong",
            "default": false
          },
          "confirmation": {
            "type": "string",
            "maxLength": 80,
            "title": "Confirmation",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApprovalRequest"
      },
...
      "ExecuteResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "queued": {
            "type": "boolean",
            "title": "Queued"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "operation_id",
          "queued"
        ],
        "title": "ExecuteResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}